requirements.txt
streamlit==1.32.0
openai==1.14.0
audio-recorder-streamlit==0.0.10
SpeechRecognition==3.10.1
soundfile==0.12.1
python-dotenv==1.0.1
